Startup works properly at last...
Turns out that you have to run inside a EscalatedSecurityContext to be able to access UserService and other services.
This is far from obvious... the docs suck big time !
Additonally there's a Timer Task atm running after one second to allow AO to activate.
It's run in a separate thread and there is no hook.
At the moment I don't want to move that into some kind of lazy loaded action.
It would make the user wait until the plugin re-indexes user repos...